chnage class names and distibute classes
