+++ automate-mautic.php
@@ -11,16 +11,8 @@
  * @package automateplus-mautic
  * @author Brainstorm Force
  */
-
 // exit if accessed directly.
 if ( ! defined( 'ABSPATH' ) ) {
 	exit();
 }
-define( 'AUTOMATEPLUS_MAUTIC_PLUGIN_DIR', plugin_dir_path( __FILE__ ) );
-define( 'AUTOMATEPLUS_MAUTIC_PLUGIN_URL', plugins_url( '/', __FILE__ ) );
-
-/**
- * Initiate plugin
- */
-require_once( AUTOMATEPLUS_MAUTIC_PLUGIN_DIR . '/classes/class-apm-automateplus.php' );
-$automate_mautic = AutomatePlus_Mautic::instance();
+require_once 'classes/class-amp-mautic-loader.php';
